feat: invite community ui component
