+++ wondrous-app/components/OnboardingDaoForm/InviteCommunity/component.tsx
@@ -1,7 +1,4 @@
-import AsanaIcon from 'components/Icons/asana.svg';
 import CSVIcon from 'components/Icons/csv.svg';
-import GithubIcon from 'components/Icons/github.svg';
-import NotionIcon from 'components/Icons/notion.svg';
 import ButtonImport from 'components/OnboardingDaoForm/ButtonImport';
 import { ButtonCSVTemplate, ButtonsWrapper, ImportButtonWrapper } from 'components/OnboardingDaoForm/styles';
 
@@ -11,24 +8,9 @@ const buttons = [
     Icon: CSVIcon,
     borderColor: `linear-gradient(270deg, #7427ff -5.62%, #06ffa5 103.12%)`,
   },
-  {
-    text: 'Import from Notion',
-    Icon: NotionIcon,
-    borderColor: `linear-gradient(270deg, #7427FF -5.62%, #FFFFFF 103.12%)`,
-  },
-  {
-    text: 'Import from Asana',
-    Icon: AsanaIcon,
-    borderColor: `linear-gradient(270deg, #7427FF -5.62%, #F06A6A 103.12%)`,
-  },
-  {
-    text: 'Import from Github',
-    Icon: GithubIcon,
-    borderColor: `linear-gradient(270deg, #7427FF -5.62%, #00A2FF 103.12%)`,
-  },
 ];
 
-const ImportTasks = () => {
+const InviteCommunity = () => {
   return (
     <ButtonsWrapper>
       <ImportButtonWrapper>
@@ -41,4 +23,4 @@ const ImportTasks = () => {
   );
 };
 
-export default ImportTasks;
+export default InviteCommunity;
