+++ wondrous-app/components/OnboardingDaoForm/InviteCommunity/index.tsx
@@ -1,0 +1,1 @@
+export { default } from './component';
