+++ wondrous-app/components/OnboardingDaoForm/index.tsx
@@ -3,3 +3,4 @@ export { default as AddImages } from './AddImages';
 export { default as CreateDao } from './CreateDao';
 export { default as DaoCategory } from './DaoCategory';
 export { default as ImportTasks } from './ImportTasks';
+export { default as InviteCommunity } from './InviteCommunity';
